yolo_prediction_service/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
Pillow>=8.0.0
numpy>=1.19.0
Flask>=2.0.0
opencv-python>=4.5.0